+++ aisec/runtime/models_test.go
@@ -1,4 +1,4 @@
-package management
+package runtime
 
 import (
 	"encoding/json"